Dispatch position-change event on change
Previously, position-change event was only dispatched when a
key press was used to change the position of the seek bar. When
the seekbar was in the process of being dragged, it only dispatched
an event at the beginning and end of the drag.

With this change, whenever the value of the seekbar changes, it will
dispatch the position-change event.